split out rule loader
